requirements.txt
anyio==4.2.0
arrow==1.3.0
binaryornot==0.4.4
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
exceptiongroup==1.2.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gspread==5.12.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
NorenRestApiPy @ file:///H:/project/ShoonyaApi-py-master/dist/NorenRestApiPy-0.0.22-py2.py3-none-any.whl
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
watchdog==3.0.0
websocket-client==1.7.0
websockets==11.0.3


